webrtc/requirements.txt
flask>=3.0.0
requests>=2.32.0
azure-identity>=1.19.0
websockets>=13.0
python-dotenv>=1.0.1
gunicorn>=23.0.0
